Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.109.2
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.16.0
httpcore==1.0.4
httplib2==0.31.2
httpx==0.27.0
idna==3.11
jiter==0.12.0
openai==1.12.0
proto-plus==1.27.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.3.2
python-dotenv==1.0.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.9.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.1
